device profiler split: kernel_profiler.hpp = X280 SPSC producer, kernel_profiler_push.hpp = DRAM (parked)
The X280 backend streams each RISC's markers into a per-RISC single-producer/single-consumer ring
in L1 and lets the X280 (L2CPU) drain them, instead of filling an L1 buffer and pushing it to DRAM.
The producer BLOCKS on a full ring, so the stream is lossless and flow-controlled, with no DRAM
traffic on quick_push()/finish().

kernel_profiler_push.hpp is the DRAM producer, parked. It is now upstream's current
kernel_profiler.hpp verbatim, which picks up all five upstream commits on that file since our merge
base: L1 buffer accumulate mode (#48506), basic profiler on Quasar (#49417), the
profiler_data_buffer data-layout fix (#50770), DeviceZoneScopedN on Quasar (#50900), and the
skip-finish_profiling-in-invalid-firmware-iterations fix (#51292).

ARCH DISPATCH. kernel_profiler.hpp now routes ARCH_QUASAR to the DRAM producer rather than
compiling the SPSC backend. This is not a preference, it is a requirement in three ways:
  - the SPSC backend needs an X280 consumer, and the X280 (L2CPU) is Blackhole hardware. On Quasar
    nothing would drain the rings, so the first full ring blocks the RISC forever;
  - upstream's Quasar support needs a RUNTIME myRiscID (internal_::get_hw_thread_idx()), while our
    ring geometry needs a compile-time one for constexpr TAIL_INDEX/HEAD_INDEX;
  - Quasar reads the NEO_REGS_0 wall-clock registers, not RISCV_DEBUG_REG_WALL_CLOCK_L, and has up
    to 24 processors per core against the 5 this ring layout assumes.
Four tt-2xx firmware files include this header, so without the guard our producer would have broken
upstream's Quasar profiler bringup. Both headers expose the same public macro API -- our surface is
a superset, and upstream's only extra macro, PROFILER_MAIN_SCOPE, has no callers outside the header
-- so no call site cares which backend is active.

Because Quasar no longer compiles the SPSC producer it does not need the zoneValid publish gate
either, so the tt-2xx/trisc.cc line this branch used to carry is dropped. zoneValid stays in the
seven tt-1xx firmware files that do compile the SPSC backend.
