+++ tt_metal/hw/firmware/src/tt-1xx/brisc.cc
@@ -99,6 +99,7 @@ uint8_t worker_logical_row_to_virtual_row[round_up_to_mult_of_4(noc_size_y)] __a
 #if defined(PROFILE_KERNEL)
 namespace kernel_profiler {
 uint32_t wIndex __attribute__((used));
+bool zoneValid __attribute__((used)) = true;  // SPSC publish gate; see kernel_profiler.hpp
 uint32_t stackSize __attribute__((used));
 uint32_t sums[SUM_COUNT] __attribute__((used));
 uint32_t sumIDs[SUM_COUNT] __attribute__((used));
