+++ tt_metal/hw/firmware/src/tt-1xx/idle_erisc.cc
@@ -78,6 +78,7 @@ inline void RISC_POST_HEARTBEAT(uint32_t& heartbeat) {
 #if defined(PROFILE_KERNEL)
 namespace kernel_profiler {
 uint32_t wIndex __attribute__((used));
+bool zoneValid __attribute__((used)) = true;  // SPSC publish gate; see kernel_profiler.hpp
 uint32_t stackSize __attribute__((used));
 uint32_t sums[SUM_COUNT] __attribute__((used));
 uint32_t sumIDs[SUM_COUNT] __attribute__((used));
